fix: Fix alignment error.
ref: https://github.com/go-ole/go-ole/issues/147

+++ safearray.go
@@ -14,7 +14,7 @@ type SafeArray struct {
 	FeaturesFlag uint16
 	ElementsSize uint32
 	LocksAmount  uint32
-	Data         uint32
+	Data         uintptr
 	Bounds       [16]byte
 }
